solutions/04-mcp-server/requirements.txt
google-adk[a2a]==1.14.1
google-cloud-compute==1.37.0
google-cloud-monitoring==2.27.2
google-cloud-modelarmor==0.3.0
google-cloud-aiplatform==1.55.0
